+++ src/components/subway/index.ts
@@ -1,2 +1,3 @@
 export { default as SubwayList } from './SubwayList';
 export { default as SubwayName } from './SubwayName';
+export { default as Days } from './Days';
