+++ src/hooks/index.ts
@@ -2,3 +2,4 @@ export { default as useGeolocation } from './useGeolocation';
 export { default as useMapPanning } from './useMapPanning';
 export { default as useTime } from './useTime';
 export { default as useDayOfWeek } from './useDayOfWeek';
+export { default as useCustomButton } from './useCustomButton';
